refactor: refactor bot permissions
